src/Entertainment Industry/ImprovAI/requirements.txt
# Core AI/ML Dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Audio Processing
librosa>=0.10.0
pretty_midi>=0.2.9
soundfile>=0.12.1
mido>=1.2.10

# User Interface
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
